bisect: add read_bisect_terms() to the public interface
Commit 833bd64f ("bisect: simplify the addition of new bisect terms",
29-06-2015) added a public function, read_bisect_terms(), to 'bisect.c'
which was then called by code in 'revision.c', having directly referenced
the symbol with an explicit external declaration. This causes sparse to
complain ('symbol not declared. Should it be static?').

In order to suppress the warning, move the external declaration to the
"bisect.h" header file and '#include' the header in 'revision.c'.

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>